* random.c (rand_init): Add a cast to fix clang compile error:   random.c:410:32: error: implicit conversion loses integer precision:   'size_t' (aka 'unsigned long') to 'int' [-Werror,-Wshorten-64-to-32]   This cast doesn't cause a problem because len is not bigger than   MT_MAX_STATE.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@41168 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ random.c
@@ -407,7 +407,7 @@ rand_init(struct MT *mt, VALUE vseed)
     else {
         if (sign != 2 && buf[len-1] == 1) /* remove leading-zero-guard */
             len--;
-        init_by_array(mt, buf, len);
+        init_by_array(mt, buf, (int)len);
     }
     if (buf != buf0) xfree(buf);
     return seed;
